requirements.txt
fastapi
uvicorn[standard]
langchain-community
youtube-transcript-api
requests
supabase
python-dotenv
psycopg2-binary
pgvector
sentence-transformers
alembic
sqlalchemy
typer
